Add files via upload
So it's meant to look like a keyhole, but its formed our of a stylised S on top of an A, Solana Auth innit...

Solana colours, transparent, that font beeman likes, etc.

Can supply as vector with layers and whatnot. I don't really know how figma works. Or github for that matter. Sorry.